fix: recover original .npmbundlerrc after failures (#71)
Unlike what I did with Babel in #82, just add a `try`/`finally` wrapper.
I wanted something more generic and reusable for Babel because we use it
in 3 places, but we only use liferay-npm-bundler in one place, so
something simpler and more direct is called for.

Test plan: install in liferay-portal, add a `throw` after bundling, and
run the `yarn build` in a module with an .npmbundlerrc file (eg.
frontend-js-web). See the original config file is restored despite the
error.

Closes: https://github.com/liferay/liferay-npm-tools/issues/71