feat: enforce canonical preset and plugin names (#89)
As described in the related issue, Babel has various optional shorthand
conventions for identifying plugins and presets, which means that it is
possible to refer to the same module under different names. Babel will
then complain about duplicates with an error message which isn't very
helpful:

    Error: Duplicate plugin/preset detected.
    If you'd like to use two separate instances of a plugin,
    they need separate names, e.g.

      plugins: [
        ['some-plugin', {}],
        ['some-plugin', {}, 'some unique name'],
      ]

It doesn't mention the actual names that are duplicated, or even if it's
a preset or a plugin that it's complaining about. This, combined with
the fact that we dynamically generate a ".babelrc" at runtime, means that
this is very confusing for people not familiar with the tooling; they
may look at their on-disk ".babelrc" and "package.json" wonder what
duplicates it is talking about.

We could silently normalize the names and merge them, but I think it's
better to do what we do in this commit: normalize to detect duplicates
and issue an error if one is found; that way, we avoid redundant config
from finding its way into the repo.

Closes: https://github.com/liferay/liferay-npm-tools/issues/89