refactor: teach getMergedConfig() and index to reject invalid types
Based on discussion about use of `console.log` in:

https://github.com/liferay/liferay-npm-tools/pull/91

I did a quick audit and found that there are only two other places in
this package where we use `console.log` and both should probably throw
errors instead of just logging a message and proceeding.

For example, current behavior:

    $ liferay-npm-scripts formatz
    'formatz' is not a valid command for liferay-npm-scripts.
    $ liferay-npm-scripts
    'undefined' is not a valid command for liferay-npm-scripts.

(Both exit with status code 0.)

Behavior after this commit:

    $ liferay-npm-scripts formatz
    /path/to/index.js:46
                    throw new Error(
                    ^

    Error: liferay-npm-scripts requires a valid command (build, format, lint, test, theme, webpack)
    $ liferay-npm-scripts
    /path/to/index.js:46
                    throw new Error(
                    ^

    Error: liferay-npm-scripts requires a valid command (build, format, lint, test, theme, webpack)

(Both exit with status code 1.)

I changed a `switch` to an object so that I could auto-generate the list
of commands, and also deleted some dead code on the way; we were
creating a temporary dir but never using it.

In any case, now we have exactly one place in the code base that uses
`console.log`.