fix: make lint/format behavior match documentation
Our documentation said that you could provide "lint" and "format"
configuration to control which paths get linted/formatted, but in
practice we were always using the "lint" config for both.

Obviously, it doesn't make a whole lot of sense to *check* formatting on
one set of globs and *apply* formatting on a different set of globs,
but in the future we're going to use ESLint in liferay-portal, at which
point "linting" and "formatting" actually will come to mean different
things. For example, we will want to "lint" only JavaScript (with
ESLint), but we'll want to "format" JS, CSS etc (with Prettier).

So, even though right now both "linting" and "formatting" are being
handled by Prettier, this change prepares us for a future in which we'll
actually have two separate tools in play.