doc: add file-naming guidance
This doesn't cover all the things yet (for example, it only talks about
".js" and ".ts"; it doesn't cover Sass or CSS or questions of folder
structure, other than mentioning "src/" and "test/"), but it's a start.

Also, it is somewhat aspirational and recommendation-al (not an actual
word) in nature, because it describes not how things are, but how I am
suggesting we should work towards in the future; concretely, reality
differs from what I've said here in three ways:

1. In liferay-portal, we currently use snake_case for files that export
   collections of things, and CamelCase for files that export
   constructors, classes, components etc.
2. In the Clay "develop" branch, we are using ".tsx" as an extension but
   I'm recommending ".ts" instead.
3. In some other repos on GitHub, we are using kebab-case in some
   filenames (eg. liferay-npm-tools).

So this is just to kick off discussion.