doc: add initial contributing guidelines
For now, just guidelines on commit messages (and effectively the same
guidance applies to PR messages). Later, we will want more
subcategories, I think, so I've split this into a table-of-contents in
CONTRIBUTING.md, then the actual content in a directory. If that turns
out to be premature hiearchical-ization the we can always fold it back
into one document later.

The idea is that a repo like Clay can just say, "For guidance on how to
format commit messages in this repo, please see ...." and link to the
commit messages doc over here (once/if we make this repo public — see:
https://github.com/liferay/liferay-frontend-guidelines/issues/5).