fix(tasks): non-CSS/JS file changes break CSS when using "watch" (#340)
For non-CSS/JS changes, we call the "deploy" task, and the first thing
that does is "build", and the first thing that does is "clean", which
deletes the CSS.

We then had a short-circuit that skipped over compiling SCSS to CSS if
the file that triggered the change wasn't a CSS file. Then when
"watch:reload" reloads the window, we can't deliver the CSS (it was
deleted and not rebuilt), so we send a 404 instead and everything is
horribly broken.

The fix then is to remove the shortcircuit, because it is better to be a
slower-but-working than faster-and-broken.

I kept the old task around even though it reduces to an empty layer of
indirection, in the name of backwards compatibility.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/340