docs: indicate the use of `--tag` when publishing
Because of the way NPM works, only one package will ever get the
"latest" tag in the registry, which leads to all sorts of confusing
behavior. `npm install` will grab the latest (most recently published),
rather than the highest semver-satisfying version. So, if we publish
9.1.4, and then 8.0.11, `npm install` will grab the 8.x release.

So, to ensure that 9.x releases are always the "latest", we need to use
tags for 8.x releases.