chore: remove sort-keys dependency
Just to reduce our dependency footprint a little and remove some
"clutter" from the source. We added this in 08fa78d03ebd197 but as far
as I can tell there is no need for the keys to be sorted (the merged
configs are never committed to the repo or inspected by a human, and we
don't do anything order-sensitive with them either).

Note that we still have another copy of sort-keys in the yarn.lock, but
that one is a transitive dependency so I'm going to ignore it for now.