feat(scripts)!: require that globs be simple (#109)
BREAKING CHANGE: In order to assign checking/fixing chores to different
underlying tools, we need to be able to split the "check" and "fix"
globs according to file type. For example, Prettier can format JS and
ESLint can check for other JS problems, but we need different tools for
JSP etc.

Rather than trying to replicate the logic that all of these tools use to
explode their globs, we just require that then globs be "simple" — ie.
non-compound. So, "*.{css,js,scss}" is bad, but "*.css", "*.js",
"*.scss" are ok. That way we can trivially assign them to the right
tools.

Why don't we want to replicate the logic? Because we'd need to find the
"lowest common denominator" which they all support, and that would be
hard to do (and not robust in the face of possible future changes in
those tools). As an example, Prettier uses "globby" for globs, which in
turn uses "fast-glob", which in turn uses "glob-parent", "is-glob", and
"micromatch"; those in turn have their own dependencies, such as
"is-extglob", "picomatch", "braces" and so on...