refactor: teach getMergedConfig() to pluck a specific property (#109)
The goal here is to make it less error-prone when accessing a specific
property of a config object. At the moment, the only time we do that is
is with "npmscripts" config, which we consume internally; all the other
configs (Babel, Jest etc) just get passed on to the external tool.

In the case of "npmscripts", we access "check" and "fix" properties in
most places, plus "preset" in one other.

To make it less likely that we pass a bad name, this commit adds an API
that performs a check that the property really exists:

    // ie:
    getMergedConfig('npmscripts', 'check');

    // instead of:
    getMergedConfig('npmscripts').check;

Obviously, nothing is to stop somebody from providing a user config with
a misspelled property like "chek", and then if we did
`getMergedConfig('npmscripts', 'chek')` we'd happily return that bogus
config without throwing an error, but the chances of two errors like
that cancelling each other out seem vanishingly small.