Fixes #345: Move default logic to config side
Instead of leaving the default value handling to the caller we do it
inside the function.

This is because all other methods are based on this pattern. Also,
because in the docs we state the default values so it makes sense
that the configuration handling code takes care of this itself to
avoid duplications.