feat(scripts): force use of local eslint-config-liferay
In practice, "eslint-config-liferay" should be available in
liferay-portal's top-level modules/node_modules folder, which means that
ESLint will be able to find it; nevertheless, just in case we ever get a
duplicate in there for whatever reason, let's make sure the one that is
bundled with liferay-npm-scripts is the one that gets used no matter
what.

The "cosmiconfig" package should have no trouble finding our base config
because it will look for "eslint.config.js" just as it looked for
"eslint.json".