chore: update and lock executable dependencies
Specifically "eslint" and "prettier". Because of the change we
implemented in 40b885590eb5dc1539 ("fix: improve executable selection on
Windows", April 2019), `spawnSync()` will preferentially try to use the
locally installed executables. We still need to be wary of duplicates
in liferay-portal though because, if I recall correctly, if there are
multiple versions of a package within a set of workspaces, the one which
"wins" and gets its executable linked in the top-level
"node_modules/.bin" folder is probably not well-defined.