feat(scripts): support multiple .eslintignore files
By design, apparently, ESLint won't let you have more than one
.eslintignore file. It will only look in the current directory, unless
you pass an explicit `--ignore-path` option. If you do that, or it finds
an .eslintignore in the current directory, it will skip over any local
"eslintIgnore" property that might exist in the package.json too.

However, we kind of need this ability because we run
`liferay-npm-scripts lint` in two modes:

1. From the "modules/" root, where we lint everything in the repo.
2. From individual subproject directories, like
   "modules/apps/segments/segments-web".

In both cases, we want to provide some very broad config that applies to
the whole repo -- eg. ignore all "*.js" files but lint all "*.es.js"
ones; for more context on that see:

https://issues.liferay.com/browse/LPS-97079

and related tickets.

But then we want to provide more granular overrides at the per-project
level so that we can slowly tighten the screws and increase our lint
coverage of the entire repo.

This is an inglorious hack, but hopefully the comments and code explain
it sufficiently.