feat: don't swallow spawn.sync errors silently (#365)
One of the things that made the actual cause of the problems on Windows
hard to understand was that we weren't printing any error messages in
the event that something went wrong with our `npm` command invocations.

So, add a wrapper that complains loudly and in detail, then aborts if
something goes wrong.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/365