fix: silent `npm install` failures on Windows in extend and kickstart
Based on my experience in #367, I searched for other usages of
`child_process` which might not work on Windows and found these two. I
believe these two `npm install` were failing on Windows as a result.

Note that we have tests asserting that errors are emitted, but they were
bad, so I removed them:

1. The tests shouldn't be actually running `npm install`; they should be
   using stubs.
2. However, in the implementation, errors are ignored. It seems pointless
   to verify that an error is emitted if you are going to ignore it
   anyway.

Also note the useless `hideOutput` flag that nobody is ever passing.

This is the 9.x port of the 8.x version of this fix (#368).

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/365