feat: make `yarn updatePackageVersions` change all usages
Previously we still had to do a manual update. I can't remember why
right now; maybe it was a misgiving about using a potentially
brittle pattern -- in any case, the whole thing is brittle so we may as
well go for it and fully automate it. If it breaks, it will break.