fix: correctly detect Windows on Travis CI
I had the `export` in there to confirm that my detection method was
correct. It was not: the `WINDIR` variable that I'd seen on the physical
Windows laptop that we have here is not set in CI.

But "OS" is (only on the Windows build, not on the Linux one).

- Sample Linux build: https://travis-ci.org/liferay/liferay-npm-tools/jobs/555683060
- Sample Windows build: https://travis-ci.org/liferay/liferay-npm-tools/jobs/555683042