fix: don't allow exploding format checks to derail CI on Windows
We don't need to run the formatting and linting checks on Windows
because those will already happen on Linux; let's just run the tests. In
this way we avoid CI blowing up on Windows like this:

    $ prettier --list-different '**/*.js' '**/*.json' '**/*.md'
    The command "yarn ci" exited with 2.
    Done. Your build exited with 1.
    [error] No matching files. Patterns tried: '**/*.js' '**/*.json' '**/*.md' !**/node_modules/** !./node_modules/** !**/.{git,svn,hg}/** !./.{git,svn,hg}/**

Evidently it is freaking out about those glob patterns due to the use of
POSIX path separators.

For consistency inside the `.travis.yml` file I just moved everything
"script-ish" out of it into separate files.