fix(scripts): don't let Windows test runs hang in CI
Our CI runs on Windows are hanging for 10 minutes and then dying:

https://travis-ci.org/liferay/liferay-npm-tools/jobs/555744779

As described here:

- https://github.com/yarnpkg/yarn/issues/2809
- https://github.com/npm/npm/issues/2938

`yarn` (and `npm`) are implemented on Windows as ".bat" scripts, and as
one of the comments on one of those issues says:

> It's actually just the ridiculous nature of DOS batch files. When a
> batch file is invoked from within another batch file, the only way to
> have execution return to the original batch files is to prefix the
> execution of the child batch file with "call".

Workaround is to call `yarn test` for each package, which means we no
longer have a .bat script calling a .bat script. We use `find` + `xargs`
because `find` and `-exec` alone won't exit with a non-zero status code
if one of the test suites fails. The nice thing about `xargs` is that it
will still run the suites for all the packages even if an earlier suite
fails, but it will reliably report the failure at the end.