test: run tests using `jest` directly in Windows
So even the trick I tried in the last commit wasn't enough. `yarn test`
bails early too because that runs `jest`, and `jest` is a .bat script,
so we're back to having one `.bat` script invoke another and failing to
wait for it.

So, we run all the tests from the top-level at once. Changed the
top-level `test` definition as well, for consistency.

Example bad run:

https://travis-ci.org/liferay/liferay-npm-tools/jobs/555762560