test: run tests using `node` directly in Windows
So that last one didn't work. Try dropping down yet another level. It
seems that on Windows, "jest" is a "/bin/sh" script that uses node to
run jest.js (there is also a ".cmd" script and a ".ps1" script, but
those don't apply in the Git Bash shell that Travis is using, as far as
I can tell).