feat(scripts): work around DDM's non-standard package.json "main" entries
I'll open a separate ticket with the folks working on DDM to address the
root cause here, but we need a workaround in the meantime to make Jest
work in their projects which have "main" entries in their "package.json"
files like "./". These projects have no "index.js" file, but ideally
they would have. For now, the projects are reaching into the directory
structures of their dependencies with full paths.

With this change, imports of both these forms will work:

    import {debounce} from 'frontend-js-web';
    import * as FormSupport from 'dynamic-data-mapping-form-renderer/js/components/FormRenderer/FormSupport.es';

Once DDM adds an actual main entry file to their projects, we'll be able
to remove this hack.