feat(scripts): add ability to lint/format changed files only
We think that the way we'll be getting passed the changed file list from
the Java build tools in liferay-portal is going to be via an environment
variable, so that's what is implemented here. Specifically, a
comma-separated list passed via LIFERAY_NPM_SCRIPTS_CHANGED_FILES.

We chose this because we can read it conveniently from inside
liferay-npm-scripts without having to pass the information down through
multiple levels. Additionally, the length limits seem to be more
liberal for environment variables than command line argument lists.

- On Windows, maximum environment variable size is either 31,767 or
  8,191 characters, but on Linux is is effectively unlimited:

  https://stackoverflow.com/questions/1078031/what-is-the-maximum-size-of-an-environment-variable-value

- On Windows, the maximum command line string is supposedly 8,191
  characters:

  https://stackoverflow.com/questions/3205027/maximum-length-of-command-line-string

  On other platforms, it is configurable but likely much higher:

  https://stackoverflow.com/questions/19354870/bash-command-line-and-input-limit

Using this command to get a crude estimate of path lengths in
liferay-portal:

    find modules/apps -name '*.es.js' \
      -not -path '*/node_modules/*' \
      -not -path '*/classes/*' \
      -not -path '*/build/*' | \
    awk '{
      len=length($0);
      total+=len
    } END {
      printf("average: %d\n", total/NR);
    }'

that shows that the average path length for ".es.js" files is 120
characters, which means that if a limit of 8,191 applied then we could
only expect to handle about 68 paths in the environment payload (on
Windows), but wouldn't have to worry about it on Linux/macOS.

If this approach ends up not working out, we can either pass the
upstream branch name instead and just compute the changed file list
internally using Git commands, or we could pass the file list in via
standard-input.

No matter what approach we end up using, we should be able to easily
adapt the `filterChangedFiles()` function to work.

When running outside of the Java tooling, `yarn format` and `yarn
checkFormat` will continue to work as before, operating without any
filtering in place.