docs: add recommendations for testing UI components
As evidenced by this mega thread:

https://github.com/jbalsas/liferay-portal/pull/1819#discussion_r303149958

This is a very complex topic with lots of tradeoffs and historical
context to weigh up. As such, this PR seeks to make some modest progress
by starting with some minimal guidelines. We begin with the bits that
are clear and uncontroversial. As we move deeper into the topic, things
become more complicated and harder to make recommendations on (and we
probably get diminishing returns anyway).