chore(preset): use loose version requirements to facilitate backporting
To make it easier to backport liferay-npm-scripts to v7.1.x of
liferay-portal, make the version requirements specified in the bundler
preset super loose: all that we really care about is that "some version
of frontend-js-web and frontend-taglib" are available, it doesn't matter
which.

Test plan: `yarn add` my local changes in liferay-portal and deploy
"fragment-web". See it still work in "master".