feat: add "postversion" scripts to automate publishing
This implements the last of the ideas in #50. (Previously we implemented
a "preversion" script and ".yarnrc" files to create the appropriate
release commits and tags.)

We add a "postversion" script that basically does this:

    # Check you are on the master branch
    git rev-parse --abbrev-ref HEAD

    # Check worktree is clean
    git diff --quite

    # Update upstream "master"
    git push upstream master

    # Merge "master" into "stable"
    git checkout stable
    git merge --ff-only master

    # Update upstream "stable"
    git push upstream stable --follow-tags

    # Return to the "master" branch
    git checkout master

    # Actually publish
    yarn publish

Obviously with abundant checks and guards and prompts along the way.

Test plan: I've sort-of-kind-of tested this as best I can without
actually doing a release (ie. by commenting out stuff, or adding
`--dry-run` to the `push` invocations, or by not supplying a OTP token
to `yarn publish` etc). It might work. Maybe.

Closes: https://github.com/liferay/liferay-npm-tools/issues/50