fix: Revert "refactor(scripts): simplify Jest config set-up" (#176)
Reported here:

https://github.com/liferay/liferay-npm-tools/issues/176

On Windows, test runs fail:

    $ yarn test
    yarn run v1.15.2
    $ liferay-npm-scripts test
    The system cannot find the file specified.
    Command L:\repo\master-portal-dev\modules\node_modules\liferay-npm-scripts\node_modules\.bin\jest --config {...} exited with code 1
    error Command failed with exit code 1.

(Edited for brevity: the config elided with "..." is about 2k
characters long.)

Seems that we are running into the limit described here:

https://devblogs.microsoft.com/oldnewthing/?p=41553

> If you are using the ShellExecute/Ex function, then you become subject
> to the INTERNET_MAX_URL_LENGTH (around 2048) command line length limit
> imposed by the ShellExecute/Ex functions.

So, let's just revert the commit and go back to how we were doing it
before. I was able to repro the failure and also confirm that reverting
fixes the problem. The contents of the config itself are fine; it's just
the length that is an issue.

This reverts commit 01932ef4c0531ca58d2c474bf8d6de490cce9020.

Closes: https://github.com/liferay/liferay-npm-tools/issues/176