chore: update to ESLint v6
In order to remain compatible with v6, we have to change our
monkey-patching strategy. Older versions of ESLint just used
`require()`, but new versions use a custom resolver.

We patch both because I don't want to have to update all the projects
that use eslint-config-liferay to ESLint v6 at the same time; they can
update to the new versions of eslint-config-liferay and eslint
separately or together. In the future once everything is on v6 we can
remove the old patching code.

Test plan: `yarn ci`, and also install this package and ESLint v6.10 in
liferay-portal and confirm that it all still works.