doc: remove mention of storybook in top-level README.md for now
We want to do a bit of "private beta testing" of the storybook feature
before publicizing it, so let's keep it out of the README for now. We
can revert this commit once we are ready for broader use.

This is a partial revert of c5433d192b278f00e5307ac2cb2ea80a551990f8.