refactor: move "stories" to "test/stories"
Checked with @jbalsas over this one: we can't really create another
top-level folder in project roots at this time, so move from:

    stories/index.js

to:

    test/stories/index.es.js

Using ".es.js" so that these files get covered by ESLint too, and I
tweaked the example in the README to ensure that if somebody copies such
an example into a file in the repo it will lint cleanly (other than the
missing copyright header).

We ruled out using "src/stories" to avoid adding stuff to the JAR files
during build (which we'd then have to exclude).

Using "test" does requires us to ignore the "stories" directory in the
test runs. I confirmed that all tests still pass by running this tool:

https://github.com/wincent/wincent/blob/7b5ee7773f47be7925c5/roles/dotfiles/files/.zsh/host/LFRW044-bin/portool-helpers/test#L13