fix: make babel-jest dependency explicit
With the update to storybook, our dependencies changed in the yarn.lock
file. When the new version was deployed to liferay-portal here:

https://github.com/brianchandotcom/liferay-portal/pull/76714

something changed in the resolution that meant that babel-jest got
installed inside node_modules/jest-config/node_modules instead of being
hoisted to the top level.

I ran the tests locally and they worked only because of a rogue copy of
babel-jest that I had in liferay-portal/../node_modules(!!) because I
must have accidentally run a `yarn` command there at some point in
ancient history.

Test plan: `yarn add` this to liferay-portal; confirm that `babel-jest`
is again in the top-level.