Fixes #303: Add a new copy root package step
In order to be able to get rid of pre-bundler build steps we are adding
a new step to the bundler that copies designated source files to the
build directory.

The idea behind that is to post-process those files with a loader after
they are copied. That way, we may run tools like Babel from the bundler
itself and/or generate JS modules from static resources (like PNGs,
CSSs, and so on).

Also, we will run rules for packages inside node_modules, so that we may
process them too as some frameworks expect. For instance, polymers ships
ES6+ code that is transpiled when building the project to optimize the
ES5 generation based on which platforms are targeted.