Fixes #347: Create a tool to link local JS Toolkit packages from yarn
The link-js-toolkit can be installed system wide by running `npm link`
inside the project. That will create a link-js-toolkit command that may be
run in yarn projects so that any JS Toolkit dependency is taken from the
local drive.

Keep in mind that you must also run `yarn link` in the Liferay JS
Toolkit projects too so that they are made available to link-js-toolkit.
Fortunately this has to be done just once and lasts forever.