Fixes #303: Change loaders API semantics
We will rename `artifacts` to `extraArtifacts` to clarify that it is
only used for new files in addition to the processed one.

Also, any loader returning `undefined` will cause the content to be left
untouched. That way, loaders that only create extra files don't need to
remember to return the original content.