Fixes #303: Let user specify complex globs in sources
We need to skip node_modules in root project so we turn sources globs
into an array and because of that, we propagate the enhancement to the
config. That way users may define complex glob expressions to decide
which rules to apply to different subsets of files in the project.